+++ odoo/local-src/bso_dealsheet/models/__init__.py
@@ -1,0 +1,7 @@
+from . import sale_dealsheet
+from . import sale_dealsheet_line
+from . import sale_dealsheet_summary_line
+from . import sale_dealsheet_wizard_confirm
+from . import sale_dealsheet_wizard_refuse
+from . import sale_dealsheet_wizard_request
+from . import sale_order